feat: add navbar and authbuttons
