+++ client/src/components/Navbar/Navbar.module.css
@@ -1,0 +1,22 @@
+.navbar {
+    background-color: #1c1c1e;  /* Dark gray */
+    padding: 15px 20px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.logo {
+    color: white;
+    font-size: 24px;
+    font-weight: bold;
+}
+
+.logout {
+    background: red;
+    color: white;
+    border: none;
+    padding: 8px 15px;
+    border-radius: 8px;
+    cursor: pointer;
+}
