+++ client/src/pages/HomePage/index.tsx
@@ -1,19 +1,13 @@
-import AuthButtons from '../../components/AuthButtons';
-import { useAuthActions } from '../../hooks/useAuthActions';
+import Navbar from '../../components/Navbar/Navbar';
 
 const HomePage = () => {
-    const { isAuthenticated, user, login, signup, logoutUser } = useAuthActions();
+
 
     return (
         <div>
-            {!isAuthenticated ? (
-                <AuthButtons login={login} signup={signup} />
-            ) : (
-                <div>
-                    <h2>Welcome, {user?.name}</h2>
-                    <button onClick={logoutUser}>Log Out</button>
-                </div>
-            )}
+            <Navbar />
+            <h2>Welcome to Fixer Finder!</h2>
+
         </div>
     )
 }
